refactor(docs): share external URL sanitizer
