+++ docs/.vitepress/data/plugins.data.ts
@@ -1,4 +1,5 @@
 import { PLUGIN_BLACKLIST } from './plugin-blacklist'
+import { sanitizeExternalUrl } from '../utils/sanitizeExternalUrl'
 
 export interface CommunityPlugin {
   name: string
@@ -8,26 +9,6 @@ export interface CommunityPlugin {
   date: string
   links: { npm: string; homepage?: string; repository?: string }
   searchText: string
-}
-
-function sanitizeExternalUrl(value: unknown): string | undefined {
-  if (typeof value !== 'string') return undefined
-
-  const trimmed = value.trim()
-  if (!trimmed) return undefined
-
-  const normalized = trimmed.startsWith('git+') ? trimmed.slice(4) : trimmed
-
-  try {
-    const parsed = new URL(normalized)
-    if (parsed.protocol === 'http:' || parsed.protocol === 'https:') {
-      return parsed.toString()
-    }
-  } catch {
-    return undefined
-  }
-
-  return undefined
 }
 
 export default {
